minor fixes for 64bit systems using collection with more than 2^32 items
git-svn-id: svn+ssh://svn.gna.org/svn/gnustep/libs/base/trunk@38295 72102866-910b-0410-8b05-ffd578937521

+++ Source/NSKeyedArchiver.m
@@ -752,12 +752,12 @@ static NSDictionary *makeReference(unsigned ref)
 	return;
 
       case _C_INT:
-	o = [NSNumber numberWithInt: *(NSInteger*)address];
+	o = [NSNumber numberWithInt: *(int*)address];
 	[_enc setObject: o forKey: aKey];
 	return;
 
       case _C_UINT:
-	o = [NSNumber numberWithUnsignedInt: *(NSUInteger*)address];
+	o = [NSNumber numberWithUnsignedInt: *(unsigned int*)address];
 	[_enc setObject: o forKey: aKey];
 	return;
